+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/Route.kt
@@ -96,4 +96,7 @@ sealed class Route {
 
 	@Serializable
 	data object Setup : Route()
+
+	@Serializable
+	data object Profiles : Route()
 }